requirements.txt
python-dotenv>=1.0.0
argparse
json5
groq